+++ TODO.MD
@@ -1,0 +1,2 @@
+[] make test for functions
+[] switch to typescript
