+++ client/src/components/DailyValueLogs/HoursWeek.js
@@ -1,0 +1,3 @@
+export default function HoursForTheWeek({ hoursWorkedWeek }) {
+  return <h2>Weekly Hours : {hoursWorkedWeek}</h2>;
+}
