+++ client/src/helpers/weeklyLogs.js
@@ -1,0 +1,33 @@
+const activeDays = [
+  {
+    id: 100,
+    day: "monday",
+    timeLog: { clockIn: "08:00", clockOut: "16:30" },
+    active: true,
+  },
+  {
+    id: 200,
+    day: "tuesday",
+    timeLog: { clockIn: "08:00", clockOut: "16:30" },
+    active: true,
+  },
+  {
+    id: 300,
+    day: "wednesday",
+    timeLog: { clockIn: "08:00", clockOut: "16:30" },
+    active: true,
+  },
+  {
+    id: 400,
+    day: "thursday",
+    timeLog: { clockIn: "08:00", clockOut: "16:30" },
+    active: true,
+  },
+  {
+    id: 500,
+    day: "friday",
+    timeLog: { clockIn: "08:00", clockOut: "16:30" },
+    active: true,
+  },
+];
+export { activeDays };
